Fix bug in PHAL::ScatterSideEqnResidual
The evaluator "diagonalizes" the residual (meaning that it
sets res=0 (and jac=1) for nodes *not* on the sideset where
the eqn is). Before, some entris that *are* on the sideset were
still diagonalized, b/c the logic detecting nodes on the side
was flawed. This commit fixes it, by gathering *all* the gids
of all the nodes on the sideset (on the current rank) first,
from which it creates the list of nodes to *not* diagonalize.