Improve how we handle configuration of Albany_config.h.in
First run configure_file into a temp file. If the destination
Albany_config.h is not present or if it's different from the
temp file, move the temp file to Albany_config.h.

This allows to avoid a rebuild of the whole project when cmake
is run, cause an Albany_config.h with a newer time stamp than
the one used in the last build would trigger a rebuild of
virtually all Albany.